+++ finalynx/portfolio/envelope.py
@@ -11,6 +11,9 @@ from .line import Line
 
 
 class EnvelopeState(Enum):
+    """Depending on the date, an envelope can be at a certain state
+    (locked, heavily taxed, free, etc)."""
+
     UNKNOWN = "Unknown"
     CLOSED = "Closed"
     LOCKED = "Locked"
@@ -19,6 +22,8 @@ class EnvelopeState(Enum):
 
 
 class Envelope:
+    """Represents an investment envelope or account (e.g. PEA, Assurance Vie, ...)."""
+
     def __init__(
         self,
         name: str,
@@ -26,12 +31,31 @@ class Envelope:
         date_created: date,
         date_unlock: Optional[date] = None,
         date_untax: Optional[date] = None,
+        key: Optional[str] = None,
     ):
+        """
+        This class represents any investment envelope or account (e.g. PEA, Assurance Vie, CTO, ...)
+        and can be used to
+
+        :param name: The envelope name (can also act as the `key` if the key is not set), will be
+        displayed in the console and webdashboard
+        :param code: The envelope name in a short version (recommended 3 characters), will be displayed
+        when rendering each line to indicate each line's envelope.
+        :param date_created: Accepts a `date` object to indicate when this envelope was opened.
+        :param date_unlock: Accepts a `date` object to indicate when this envelope will be unlocked. For
+        instance, a PEA will be unlocked 5 years after creation as required by law.
+        :param date_untax: Accepts a `date` object to indicate when this envelope will no longer have
+        higher-than-usual tax rates. For instance, a French Assurance Vie is highly taxed for the first
+        8 years after creation date, and then changes to a better tax rate.
+        :param key: If you want to set a different name here from Finary's name, set the Finary name in
+        this `key` parameter. The field `name` will be used when displaying the envelope in Finalynx.
+        """
         self.name = name
         self.code = code
         self.date_created = date_created
         self.date_unlock = date_unlock if date_unlock else date_created
         self.date_untax = date_untax if date_untax else date_created
+        self.key = key
 
         if not (self.date_created <= self.date_unlock <= self.date_untax):
             raise ValueError("Envelope dates must be ordered by created <= unlock <= untax")
@@ -75,24 +99,64 @@ class Envelope:
 
 
 class PEA(Envelope):
-    def __init__(self, name: str, code: str, date_created: date):
+    """Handy shortcut to quickly define a PEA (automatically sets 5-years locked)."""
+
+    def __init__(self, name: str, code: str, date_created: date, key: Optional[str] = None):
+        """Declare a PEA (locked for 5 years) with:
+        :param name: Name of your account.
+        :param code: Short name of your account (3 characters recommended).
+        :param date_created: A `date` instance of the account creation date.
+        :param key: Optional, if you want to use a different name in Finalynx than in Finary.
+        See `Envelope`'s documentation for additional details.
+        """
         date_unlock = date_created + relativedelta(years=5)
-        super().__init__(name, code, date_created, date_unlock, date_unlock)
+        super().__init__(name, code, date_created, date_unlock, date_unlock, key=key)
 
 
 class PEE(Envelope):
-    def __init__(self, name: str, code: str, date_created: date, date_unlock: Optional[date] = None):
+    """Handy shortcut to quickly define a PEE (automatically sets 5-years locked)."""
+
+    def __init__(
+        self, name: str, code: str, date_created: date, date_unlock: Optional[date] = None, key: Optional[str] = None
+    ):
+        """Declare a PEE (locked for 5 years) with:
+        :param name: Name of your account.
+        :param code: Short name of your account (3 characters recommended).
+        :param date_created: A `date` instance of the account creation date.
+        :param date_unlock: A `date` instance of the account unlock date, defaults to 5 years.
+        :param key: Optional, if you want to use a different name in Finalynx than in Finary.
+        See `Envelope`'s documentation for additional details.
+        """
         if not date_unlock:
             date_unlock = date_created + relativedelta(years=5)
-        super().__init__(name, code, date_created, date_unlock, date_unlock)
+        super().__init__(name, code, date_created, date_unlock, date_unlock, key=key)
 
 
 class AV(Envelope):
-    def __init__(self, name: str, code: str, date_created: date):
+    """Handy shortcut to quickly define a PEA (automatically sets 8-years taxed)."""
+
+    def __init__(self, name: str, code: str, date_created: date, key: Optional[str] = None):
+        """Declare an Assurance Vie (taxed for 8 years) with:
+        :param name: Name of your account.
+        :param code: Short name of your account (3 characters recommended).
+        :param date_created: A `date` instance of the account creation date.
+        :param key: Optional, if you want to use a different name in Finalynx than in Finary.
+        See `Envelope`'s documentation for additional details.
+        """
         date_untax = date_created + relativedelta(years=8)
-        super().__init__(name, code, date_created, date_created, date_untax)
+        super().__init__(name, code, date_created, date_created, date_untax, key=key)
 
 
 class PER(Envelope):
-    def __init__(self, name: str, code: str, date_created: date, date_retirement: date):
-        super().__init__(name, code, date_created, date_retirement, date_retirement)
+    """Handy shortcut to quickly define a PER (locked until retirement)."""
+
+    def __init__(self, name: str, code: str, date_created: date, date_retirement: date, key: Optional[str] = None):
+        """Declare a PER (locked until retirement) with:
+        :param name: Name of your account.
+        :param code: Short name of your account (3 characters recommended).
+        :param date_created: A `date` instance of the account creation date.
+        :param date_retirement: A `date` instance of your expected retirement date.
+        :param key: Optional, if you want to use a different name in Finalynx than in Finary.
+        See `Envelope`'s documentation for additional details.
+        """
+        super().__init__(name, code, date_created, date_retirement, date_retirement, key=key)